+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Fixed styling for itemize.
 - Watch mode now refreshes pages after compile. 
+- The `documentclass` macro is now hidden from the main page.
 
 
 # v0.2.3, 2026-04-23