+++ changelog.md
@@ -5,6 +5,7 @@
 - The `documentclass` macro is now hidden from the main page.
 - Added the comments feature. 
 - Reorganised the usage documentation to not remain on one page. 
+- Fixed a bug where enumerates require an additional line break after. 
 
 
 # v0.2.3, 2026-04-23